Ensure we are using utf-8 for file IO
Without this on a system that does not have utf-8 as the default (e.g. the latest https://hub.docker.com/_/ruby/), the user will hit an error as soon as JSON.parse() gets to the apostrophe in the example.json.

`ERROR Encoding::InvalidByteSequenceError: "\xE2" on US-ASCII`

This fix makes sure that wen we do File.read() it is going to return utf-8 by default, which is what the script was originally expecting.